# backend/docs/swagger.yaml
openapi: "3.0.3"
info:
  title: TrustChain API
  description: >
    Blockchain-based trust framework for location recommendations.
    All responses follow a standard envelope: { success, message, data } or { success, message, error }.
  version: "1.0.0"
  contact:
    name: TrustChain Backend
    email: backend@trustchain.dev
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /pois:
    get:
      tags: [pois]
      summary: List points of interest
      operationId: getPOIs
      parameters:
        - name: lat
          in: query
          required: false
          schema:
            type: number
            format: double
            minimum: -90
            maximum: 90
        - name: lon
          in: query
          required: false
          schema:
            type: number
            format: double
            minimum: -180
            maximum: 180
        - name: radius
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50000
            default: 1000
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: POIs retrieved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/POIListResponse"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  schemas:
...
    POIListResponse:
      allOf:
        - $ref: "#/components/schemas/SuccessEnvelope"
        - type: object
          properties:
            data:
              type: object
              properties:
                count: { type: integer }
                pois:
                  type: array
                  items:
                    $ref: "#/components/schemas/POI"
...
    POI:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        category: { type: string }
        tags:
          type: array
          items: { type: string }
        location:
          $ref: "#/components/schemas/LatLon"
        address:
          type: object
          properties:
            street: { type: string }
            city: { type: string }
            country: { type: string }
            postalCode: { type: string }
        metadata:
          type: object
          properties:
            averageRating: { type: number }
            totalReviews: { type: integer }
            totalCheckins: { type: integer }
            verified: { type: boolean }
        createdAt: { type: string, format: date-time }
    LatLon:
      type: object
      properties:
        latitude: { type: number, format: double }
        longitude: { type: number, format: double }
    SuccessEnvelope:
      type: object
      required: [success, message]
      properties:
        success: { type: boolean, example: true }
        message: { type: string }
    ErrorResponse:
      type: object
      required: [success, message, error]
      properties:
        success: { type: boolean, example: false }
        message: { type: string }
        error: { type: string }
  responses:
...
    UnprocessableEntity:
      description: Semantic validation failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...